+++ inst/shiny/server.R
@@ -3,15 +3,26 @@
 shinyServer(function(input, output, session) {
   # Database of choice
   FELLA.DATA <- reactive({
-    loadKEGGdata(
-      databaseDir = input$database, 
-      internalDir = FALSE, 
-      loadMatrix = "all")
+    if (input$database != "") {
+      loadKEGGdata(
+        databaseDir = input$database, 
+        internalDir = FALSE, 
+        loadMatrix = c("diffusion", "pagerank")
+      )
+    }
   })
   # database summary
   output$databaseInfo <- renderText({
     data <- FELLA.DATA()
-    comment(FELLA:::getGraph(data))
+    if (is.null(data)) {
+        c(
+            "Database directory is empty...", 
+            "\nMake you executed:", 
+            "\nbuildGraphFromKEGGREST", 
+            "\nbuildDataFromGraph")
+    } else {
+        comment(FELLA:::getGraph(data))
+    }
   })
   
   # First step: create the USER variable
@@ -137,7 +148,34 @@ shinyServer(function(input, output, session) {
   # This function returns the currently chosen graph connected component
   currentGraph <- reactive({
     if (!is.null(createUser())) {
-      return(generateGraph())
+      g <- generateGraph()
+      # GO tag?
+      orgDb <- input$GOorgDb
+      ont <- input$GOOntology
+      biomart <- input$GObiomart
+      dataset <- input$GOdataset
+      if (input$GOTermInput != "" & 
+          input$method != "hypergeom" &
+          !is.null(orgDb) & 
+          !is.null(ont) & 
+          !is.null(biomart) & 
+          !is.null(dataset)) {
+        return(
+          addGOToGraph(
+            graph = g, 
+            GOterm = input$GOTermInput, 
+            godata.options = list(
+              OrgDb = orgDb, 
+              ont = ont
+            ), 
+            mart.options = list(
+              biomart = biomart, 
+              dataset = dataset
+            )
+          )
+        )
+      }
+      return(g)
     }
     return(NULL)
   })
@@ -157,9 +195,12 @@ shinyServer(function(input, output, session) {
   # Table of results
   output$tableSolution <- DT::renderDataTable({
     data <- FELLA.DATA()
-    if (!is.null(data)) {
+    user <- createUser()
+    g <- currentGraph()
+    if (!is.null(data) & !is.null(g) & is.igraph(g)) {
+      if (vcount(g) > 0) return(NULL)
       wholeTable <- generateResultsTable(
-        object = createUser(), 
+        object = user, 
         method = input$method, 
         threshold = input$threshold,
         plimit = 15, 
@@ -167,41 +208,33 @@ shinyServer(function(input, output, session) {
         LabelLengthAtPlot = 100, 
         data = data)
       
-      plottedRows <- wholeTable$"KEGG id" %in% V(currentGraph())$name
+      # Only nodes in graph
+      plottedRows <- wholeTable$"KEGG.id" %in% V(g)$name
       outTable <- wholeTable[plottedRows, ]
       rownames(outTable) <- NULL
       
-      DT::datatable(outTable)
+      # Add hyperlinks to KEGG
+      outTable$"KEGG.id" <- paste0(
+        "<a href=\"http://www.genome.jp/dbget-bin/www_bget?",
+        outTable$"KEGG.id", "\"", "\ target=\"_blank", 
+        "\">", outTable$"KEGG.id", "</a>")
+      escape <- which(colnames(outTable) == "KEGG.id")
+      
+      DT::datatable(outTable, escape = escape) %>%
+        DT::formatSignif(columns = "p.score", digits = 2)
     }
   })
   
   # ---------------------------------------------------
   #  CC example: change default value for the updateTextInput!
   observe({
-    if (input$exampleGOCC > 0)
+    if (input$exampleGOCC > 0) {
       updateTextInput(session, "GOTermInput", value = "GO:0005739")
-  })
-  
-  # SAVE YOUR RESULTS
-  output$saveText <- reactive({
-    if (input$saveButton) {
-      directory <- paste0(getwd(), "/", input$saveName)
-      
-      dir.create(directory)
-      
-      pdf(paste0(directory, "/", "test.pdf"), width = 16, height = 16)
-      plotPagerank()
-      dev.off()
-      
-      if (!is.null(createUser())) 
-        write.csv(
-          createUser()@pagerank@pscores,
-          file = paste0(directory, "/table.csv"))
-      
-      return(paste0("Results saved in directory ", directory))
-      
+      updateTextInput(session, "GOorgDb", value = "org.Hs.eg.db")
+      updateSelectInput(session, "GOOntology", selected = "CC")
+      updateTextInput(session, "GObiomart", value = "ensembl")
+      updateTextInput(session, "GOdataset", value = "hsapiens_gene_ensembl")
     }
-    return("Introduce the directory to save your outputs")
   })
   
   ###########################################################
@@ -213,6 +246,14 @@ shinyServer(function(input, output, session) {
       label <- V(g)$label
       nodes <- data.frame(id, label, stringsAsFactors = FALSE)
       
+      # GO labels?
+      if ("GO.simil" %in% list.vertex.attributes(g)) {
+        GO.simil <- unlist(V(g)$GO.simil)
+        GO.annot <- TRUE
+      } else {
+        GO.annot <- FALSE
+      }
+      
       map.com <- c("pathway", "module", "enzyme", "reaction", "compound")
       map.color <- c("#E6A3A3", "#E2D3E2", "#DFC1A3", "#D0E5F2", "#A4D4A4")
       map.labelcolor <- c("#CD0000", "#CD96CD", "#CE6700", "#8DB6CD", "#548B54")
@@ -223,22 +264,44 @@ shinyServer(function(input, output, session) {
         "box",
         "ellipse"
       )
-      # 
       nodes$group <- map.com[V(g)$com]
       nodes$color <- map.color[V(g)$com]
-      # nodes$nodeLabelColor <- map.labelcolor[V(g)$com]
+
       # width
       nodes$value <- map.nodeWidth[V(g)$com]
-      # nodes$height <- map.nodeWidth[V(g)$com]
       nodes$shape <- nodeShape
+      
+      # Change color and label if GO annotations are present
+      if (GO.annot) {
+        ids <- !is.na(GO.simil)
+        GO.semsim <- GO.simil[ids]
+        GO.hits <- names(GO.semsim)
+        if (!is.null(GO.hits)) {
+          newColor <- sapply(
+            GO.semsim, 
+            function(x) {
+              if (x < 0.5) return("#FFD500")
+              else if (x < 0.7) return("#FF5500")
+              else if (x < 0.9) return("#FF0000")
+              return("#B300FF")
+            }
+          )
+          newName <- paste0(nodes$label[ids], "[", GO.hits, "]")
+          newShape <- "triangle"
+          
+          # modify name and color
+          nodes$label[ids] <- newName
+          nodes$color[ids] <- newColor
+          nodes$shape[ids] <- newShape
+        }
+      }
+      
       # tooltip
       nodeLink <- paste0(
         "<a href=\"http://www.genome.jp/dbget-bin/www_bget?",
         V(g)$name, "\"", "\ target=\"_blank", "\">", V(g)$name, "</a>")
 
       nodes$title <- nodeLink
-      # nodes$x <- (plotSolution()$x)*600
-      # nodes$y <- -(plotSolution()$y)*800
       
       source <- V(g)[get.edgelist(g)[, 1]]$name
       target <- V(g)[get.edgelist(g)[, 2]]$name
@@ -354,9 +417,7 @@ shinyServer(function(input, output, session) {
           plimit = 15, 
           data = data)
       }
-      
     } 
-    #     , contentType = "text/csv"
   )
   
   #   output$exportcsv = downloadHandler(